python-bot/requirements.txt
discord.py==2.5.3
python-dotenv==1.0.0
yt_dlp==2025.10.21
PyNaCl==1.5.0
aiohttp==3.9.4
mutagen==1.46.0
pytube==14.2.0
ffmpeg-python==0.5.0
